requirements.txt
anthropic
python-dotenv
# system install portaudio
pyaudio
speechrecognition
pyttsx3
elevenlabs
pydub
